+++ packages/db/src/schema/calendar.ts
@@ -1,0 +1,19 @@
+import { type SQL, sql } from "drizzle-orm";
+import { date, pgTable, varchar } from "drizzle-orm/pg-core";
+import { term } from ".";
+
+export const calendarTerm = pgTable("calendar_term", {
+  id: varchar("id")
+    .primaryKey()
+    .generatedAlwaysAs(
+      (): SQL =>
+        sql`${calendarTerm.year} || ' ' || CASE WHEN ${calendarTerm.quarter} = 'Fall' THEN 'Fall' WHEN ${calendarTerm.quarter} = 'Winter' THEN 'Winter' WHEN ${calendarTerm.quarter} = 'Spring' THEN 'Spring' WHEN ${calendarTerm.quarter} = 'Summer1' THEN 'Summer1' WHEN ${calendarTerm.quarter} = 'Summer10wk' THEN 'Summer10wk' WHEN ${calendarTerm.quarter} = 'Summer2' THEN 'Summer2' ELSE '' END`,
+    ),
+  year: varchar("year").notNull(),
+  quarter: term("quarter").notNull(),
+  instructionStart: date("instruction_start", { mode: "date" }).notNull(),
+  instructionEnd: date("instruction_end", { mode: "date" }).notNull(),
+  finalsStart: date("finals_start", { mode: "date" }).notNull(),
+  finalsEnd: date("finals_end", { mode: "date" }).notNull(),
+  socAvailable: date("soc_available", { mode: "date" }).notNull(),
+});
