+++ packages/db/src/schema/course-materials.ts
@@ -1,5 +1,5 @@
 import { index, pgEnum, pgTable, text, uuid, varchar } from "drizzle-orm/pg-core";
-import { websocSection } from "./websoc.ts";
+import { websocSection } from ".";
 
 export const materialTerms = ["Fall", "Winter", "Spring", "Summer"] as const;
