+++ packages/db/src/schema/larc.ts
@@ -1,0 +1,26 @@
+import { boolean, index, pgTable, timestamp, uuid, varchar } from "drizzle-orm/pg-core";
+import { websocCourse } from ".";
+
+export const larcSection = pgTable(
+  "larc_section",
+  {
+    id: uuid("id").primaryKey().defaultRandom(),
+    courseId: uuid("course_id")
+      .references(() => websocCourse.id)
+      .notNull(),
+    daysString: varchar("days_string").notNull(),
+    timeString: varchar("time_string").notNull(),
+    startTime: timestamp("start_time", { mode: "date" }).notNull(),
+    endTime: timestamp("end_time", { mode: "date" }).notNull(),
+    instructor: varchar("instructor").notNull(),
+    building: varchar("building").notNull(),
+    meetsMonday: boolean("meets_monday").notNull().default(false),
+    meetsTuesday: boolean("meets_tuesday").notNull().default(false),
+    meetsWednesday: boolean("meets_wednesday").notNull().default(false),
+    meetsThursday: boolean("meets_thursday").notNull().default(false),
+    meetsFriday: boolean("meets_friday").notNull().default(false),
+    meetsSaturday: boolean("meets_saturday").notNull().default(false),
+    meetsSunday: boolean("meets_sunday").notNull().default(false),
+  },
+  (table) => [index().on(table.courseId)],
+);
